unity build for tests

+++ templates/linked_list.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include "../types.h"
 
-typedef i32 GENERIC_TYPE;
+typedef u64 GENERIC_TYPE;
 
 typedef struct Node {
   struct Node* next;
